requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.86.0
anyio==4.13.0
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
distro==1.9.0
docstring_parser==0.17.0
elevenlabs==2.40.0
ffmpeg-python==0.2.0
filelock==3.25.2
fsspec==2026.2.0
future==1.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
jiter==0.13.0
llvmlite==0.46.0
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
more-itertools==10.8.0
mpmath==1.3.0
networkx==3.6.1
numba==0.64.0
numpy==2.4.3
openai==2.30.0
openai-whisper==20250625
praw==7.8.1
prawcore==2.4.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
python-dotenv==1.2.2
regex==2026.2.28
requests==2.32.5
rich==14.3.3
schedule==1.2.2
setuptools==70.2.0
shellingham==1.5.4
sniffio==1.3.1
sympy==1.14.0
tiktoken==0.12.0
torch==2.11.0+cpu
tqdm==4.67.3
triton==3.6.0
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
update-checker==0.18.0
urllib3==2.6.3
websocket-client==1.9.0
websockets==16.0
